Refactor TUI into dedicated modules
